requirements.txt
flask
pymongo[srv]
redis
python-dotenv
pydantic
gunicorn
pyjwt
werkzeug
pytest
fakeredis
mongomock
flask_cors